requirements.txt
django
djangorestframework
djangorestframework_simplejwt
psycopg2
pilow
django-filter
markdown